+++ AddOns/HandyNotes_BattleForAzerothTreasures/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Changed in upcoming release
+
+* Options for showing/hiding points in particular zones
+* Optimize some checks by using GetPlayerAuraBySpellID
+
 ## Changed in v10
 
 * Add achievements: